Move EVM assets UI into module
