+++ index.html
@@ -65,7 +65,7 @@
     <!-- <link rel="manifest" href="./manifest.json" /> -->
   </head>
   <body>
-    <script src="./app.js?v=1455-11" type="module"></script>
+    <script src="./app.js?v=1455-12" type="module"></script>
     <script src="./lib.js" type="module"></script>
     <script src="./crypto.js" type="module"></script>
     <div class="container">
